Fix the scrim on collapsing layout view
Put the toolbar as the last element in the collapsing layout view so
that the toolbar color shows when the toolbar is collapsed.

Add a scrim view so the text shows up on collapsing layout view.

Remove the bindViews() from the onCreateView so that the text views can
show up in all of the detail layouts.

Set the detail view to invisible so we do not see the layouts render
before there is any content.